Rename lib/state to lib/store
